feat: Add FichaUsuario component and routes
- Add FichaUsuario component to display user information
- Add FichaUsuario routes for navigation

Refactor FichaUsuarioPage and FichaUsuariosLayout components

- Remove FichaUsuarioPage component
- Remove FichaUsuariosLayout component

refactor: Update Home component

- Update Home component to match FichaUsuario component structure
