+++ src/app/modules/fichausuarios/fichaUsuario.routes.tsx
@@ -1,0 +1,9 @@
+import { Route, Routes } from 'react-router-dom'
+import { FichaUsuario } from './pages/FichaUsuario'
+
+export const FichaUsuarioRoutes = () => (
+  <Routes>
+    <Route path='ficha-usuario' element={<FichaUsuario />} />
+    <Route index element={<FichaUsuario />} />
+  </Routes>
+)
