+++ src/app/modules/fichausuarios/pages/FichaUsuario.tsx
@@ -1,0 +1,7 @@
+export function FichaUsuario() {
+  return (
+    <div>
+      Ficha Usuario
+    </div>
+  )
+}
