+++ src/app/modules/home/home.routes.tsx
@@ -1,0 +1,12 @@
+import { Route, Routes } from "react-router-dom";
+import { Home } from "./pages/home";
+
+interface HomeRoutesProps { }
+
+export function HomeRoutes(props: HomeRoutesProps): JSX.Element {
+    return (
+        <Routes>
+            <Route index element={<Home />} />
+        </Routes>
+    );
+}
